+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.6.1] - 2026-04-07
 - Fixed a bug creating copy of index directory
 - Expanded history to previous sessions in REPL 
+- Added path completion for REPL commands and file paths
+- Fixed wrong information about GPU use 
 
 ## [0.6.0] - 2026-04-06
 - Added review-session export with `/export` and `--output`.